refactor(next): move course files to course dir
Also implemented button components
