+++ frontend/src/components/coursepage/Assignments.tsx
@@ -2,6 +2,7 @@
 
 import React, { useState } from "react";
 import CreateAssignment from "./CreateAssignment";
+import AddButton from "@/components/buttons/AddButton";
 
 interface Assignments {
   id: string;
@@ -18,10 +19,6 @@ const Assignments = () => {
 
   const handleCreateAssignment = (assignmentData: any) => {
     setAssignments([...assignments, assignmentData]);
-  };
-
-  const handleMakeAssignment = (e: any) => {
-    setIsCreatingAssignment(true);
   };
 
   const handleSelectChange = (event: React.ChangeEvent<HTMLSelectElement>) => {
@@ -53,12 +50,7 @@ const Assignments = () => {
     <div className="w-full">
       <div className="flex justify-between border-b-2 border-white mb-4 pb-4">
         <h1 className="text-white font-bold text-2xl">Assignments</h1>
-        <button
-          className="rounded-full bg-white text-black text-sm font-light py-1 px-2 mt-2 flex items-center justify-center"
-          onClick={handleMakeAssignment}
-        >
-          +
-        </button>
+          <AddButton onClick={() => {setIsCreatingAssignment(true);}} />
       </div>
       <select value={selectedAssignment} onChange={handleSelectChange}>
         <option value="">Choose an assignment</option>
